# Draft: ToCTree env collector: disable read-parallelism