### Feature or Bugfix
- Bugfix

### Purpose
Bugreports #6714 and #12409 indicate that the table-of-contents collection process is not currently implemented in a way that guarantees deterministic resolution in the presence of parallelism.

Until such time as we can implement that, disable parallel reads for the `TocTreeCollector` so that the tree is resolved serially.

### Detail
- Disable parallel reads for the `TocTreeCollector` by setting `parallel_read_safe: False` for it.

### Relates
- May resolve #6714.
- May resolve #12409.