@kartben 2024-09-12
Would be great to get a rough idea of the potential impact on build speed (if any). Thanks!