@chrisjsewell 2024-09-12
@jayaddison you realise that this would literally mean it would be impossible to run sphinx in parallel 😅
This is a non-starter, you need to actually fix the problem