@jayaddison 2024-09-12
> @jayaddison you realise that this would literally mean it would be impossible to run sphinx in parallel 😅 This is a non-starter, you need to actually fix the problem

Agreed; I didn't realize that disabling read-parallelism for a single extension also causes read-parallelism to be disabled for all other extensions within the same build.

Fortunately there is [at least one `testroot`](https://github.com/sphinx-doc/sphinx/issues/6714#issuecomment-2344996426) (`test-toctree-glob`) that exhibits the `toctree` parallel-build-non-determinism, so it may be possible to narrow in on the problem more easily now.