@jayaddison 2024-09-12
> I tested that this fully helps with linux kernel-docs reproducibilty.

Thank you, @bmwiedemann :+1: